Refactors things a bit
